add requests and test

+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -3,8 +3,9 @@ package ru.practicum.shareit.user.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import ru.practicum.shareit.user.model.User;
+import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.shareit.user.dto.UserDto;
+import ru.practicum.shareit.user.model.User;
 import ru.practicum.shareit.exception.user.UserNotFoundException;
 import ru.practicum.shareit.exception.user.UserWithoutEmailException;
 import ru.practicum.shareit.user.mapper.UserMapper;
@@ -15,9 +16,11 @@ import java.util.List;
 @Service
 @Slf4j
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
 
+    @Transactional
     @Override
     public UserDto create(UserDto userDto) {
         if (userDto.getEmail() == null) throw new UserWithoutEmailException("without email");
@@ -28,11 +31,11 @@ public class UserServiceImpl implements UserService {
         return UserMapper.toUserDto(newUser);
     }
 
+    @Transactional
     @Override
     public UserDto update(int userId, UserDto userDto) {
         UserDto oldUserDto = get(userId);
         checkDataForUpdate(userDto, oldUserDto);
-
         User user = userRepository.save(UserMapper.toUser(userDto));
         log.info("Update User with ID:{}", userId);
         return UserMapper.toUserDto(user);
@@ -52,10 +55,15 @@ public class UserServiceImpl implements UserService {
         return UserMapper.listToUserDto(userRepository.findAll());
     }
 
+    @Transactional
     @Override
-    public void delete(int userId) {
-        userRepository.deleteById(userId);
-        log.info("Delete User with ID:{}", userId);
+    public boolean delete(int userId) {
+        if (userRepository.existsById(userId)) {
+            userRepository.deleteById(userId);
+            log.info("Delete User with ID:{}", userId);
+            return true;
+        }
+        return false;
     }
 
     private void checkDataForUpdate(UserDto updateUser, UserDto oldUserDto) {
